+++ src/netcore/tests/task-runners/success-date-transforms.ts
@@ -1,0 +1,11 @@
+import path = require('path');
+import models = require('../builders/request-model-builder');
+
+// const taskPath = path.join(__dirname, '../dist', 'index.js');
+const taskPath = path.join('c:\\DEV\\GIT\\vsts-assemblyinfo-task\\src\\netcore\\dist', 'index.js');
+
+const model = new models.RequestModel(taskPath)
+    .withFailOnWarning(true)
+    .withCompany('Bleddyn Richards Inc $(date:YYYY)')
+    .withDescription('Assembly Info $(date:YYYY) is an extension for Azure DevOps that sets assembly information from a build.')
+    .build();